+++ NEWS.md
@@ -4,6 +4,8 @@ __New Functionality__
 
 - `tk_time_series_cv_plan()` - Now works with k-fold cross validation objects from `vfold_cv()` function. 
 
+- `step_slidify_agument()` - A variant of step slidify that adds multiple rolling columns inside of a recipe. 
+
 - `pad_by_time()` - Added new argument `.fill_na_direction` to specify a `tidyr::fill()` strategy for filling missing data. 
 
 __Bug Fixes__